% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cp_model}
\alias{cp_model}
\title{nflfastR Completion Probability Model}
\format{
A xgb.Booster
}
\source{
See \url{https://www.opensourcefootball.com/posts/2020-09-28-nflfastr-ep-wp-and-cp-models/}
for the code to build the model.
}
\usage{
cp_model
}
\description{
nflfastR Completion Probability Model
}
\keyword{datasets}
