% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ep_model}
\alias{ep_model}
\title{nflfastR Expected Points Model}
\format{
A xgb.Booster
}
\source{
See \url{https://www.opensourcefootball.com/posts/2020-09-28-nflfastr-ep-wp-and-cp-models/}
for the code to build the model.
}
\usage{
ep_model
}
\description{
nflfastR Expected Points Model
}
\keyword{datasets}
