% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fg_model}
\alias{fg_model}
\title{nflfastR Field Goal Model}
\format{
A generalized additive model (GAM)
}
\source{
See \url{https://www.opensourcefootball.com/posts/2020-09-28-nflfastr-ep-wp-and-cp-models/}
for the code to build the model.
}
\usage{
fg_model
}
\description{
nflfastR Field Goal Model
}
\keyword{datasets}
