% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wp_model_spread}
\alias{wp_model_spread}
\title{nflfastR Win Probability Model (including Spread)}
\format{
A xgb.Booster
}
\source{
See \url{https://www.opensourcefootball.com/posts/2020-09-28-nflfastr-ep-wp-and-cp-models/}
for the code to build the model.
}
\usage{
wp_model_spread
}
\description{
nflfastR Win Probability Model (including Spread)
}
\keyword{datasets}
