% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xyac_model}
\alias{xyac_model}
\title{nflfastR Expected Yards after Catch Model}
\format{
A xgb.Booster
}
\source{
See \url{https://www.opensourcefootball.com/posts/2020-09-28-nflfastr-ep-wp-and-cp-models/}
for the code to build the model.
}
\usage{
xyac_model
}
\description{
nflfastR Expected Yards after Catch Model
}
\keyword{datasets}
