% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastrmodels-package.R
\docType{package}
\name{fastrmodels-package}
\alias{fastrmodels}
\alias{fastrmodels-package}
\title{fastrmodels: Models for the 'nflfastR' Package}
\description{
A data package that hosts all models for the
    'nflfastR' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrcaseb/fastrmodels}
  \item Report bugs at \url{https://github.com/mrcaseb/fastrmodels/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

Authors:
\itemize{
  \item Ben Baldwin \email{bbaldwin206@gmail.com}
}

}
\keyword{internal}
